utils.h moved into utils/misc.h
