+++ c/frame/integrity_check.cc
@@ -15,9 +15,9 @@
 //------------------------------------------------------------------------------
 #include "frame/py_frame.h"
 #include "utils/exceptions.h"
+#include "utils/misc.h"      // repr_utf8
 #include "datatable.h"
 #include "encodings.h"
-#include "utils.h"      // repr_utf8
 
 
 //------------------------------------------------------------------------------
