+++ c/utils/misc.cc
@@ -5,7 +5,7 @@
 //
 // © H2O.ai 2018
 //------------------------------------------------------------------------------
-#include "utils.h"
+#include "utils/misc.h"
 #include <stdint.h>
