+++ c/utils/misc.h
